auth-service/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.1.8
dotenv==0.9.9
Flask==3.1.3
Flask-MySQLdb==2.0.0
gunicorn==23.0.0
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.7
packaging==26.0
PyJWT==2.12.1
python-dotenv==1.2.1
typing_extensions==4.15.0
Werkzeug==3.1.6
zipp==3.23.0
